requirements.txt
streamlit
spacy
pymupdf
pdfplumber
python-docx
nltk